Add InclusiveNamespace support to xmerl_dsig
After this, we also check the signature method and c14n method to make
sure they're the ones that we actually support (now we will crash
if they are not).

Will add some tests for this later, generating small XML to test
it is quite complicated at the moment (since they need to be signed in
the new canonical form).

Probably solves #7